+++ src/frontend/komponenter/ManuellJournalfør/Dokument/Dokumenter.tsx
@@ -4,7 +4,7 @@ import { AlertStripeAdvarsel } from 'nav-frontend-alertstriper';
 
 import { RessursStatus } from '@navikt/familie-typer';
 
-import { useManuellJournalfør } from '../../context/ManuellJournalførContext';
+import { useManuellJournalfør } from '../../../context/ManuellJournalførContext';
 import { DokumentVelger } from './DokumentVelger';
 
 export const Dokumenter: React.FC = () => {
